services: rename repoimporter module to repos
