+++ setup.py
@@ -147,7 +147,7 @@ setup(
             'cap_services = cap.modules.services.views:blueprint',
             'cap_schemas = cap.modules.schemas.views:blueprint',
             'cap_auth = cap.modules.auth.views:blueprint',
-            'cap_repos = cap.modules.repoimporter.views:repos_bp',
+            'cap_repos = cap.modules.repos.views:repos_bp',
             'invenio_oauthclient = invenio_oauthclient.views.client:blueprint',
         ],
         'invenio_celery.tasks': [
@@ -176,7 +176,7 @@ setup(
             'analysis_schema_model = cap.modules.schemas.models',
             'status_check_model = cap.modules.services.models',
             'auth = cap.modules.auth.models',
-            'git_model = cap.modules.repoimporter.models',
+            'repos_model = cap.modules.repos.models',
         ],
         'invenio_db.alembic': ['cap = cap:alembic'],
         'invenio_config.module': ['cap = cap.config']
